requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=20.1.0=py_0
blas=1.0=mkl
boto3=1.14.48=py_0
botocore=1.17.48=py_0
brotlipy=0.7.0=py38he774522_1000
ca-certificates=2020.7.22=0
catalogue=1.0.0=py38_1
certifi=2020.6.20=py38_0
cffi=1.14.2=py38h7a1dbc1_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=py_0
cryptography=3.0=py38h7a1dbc1_0
cudatoolkit=10.2.89=h74a9793_1
cymem=2.0.3=py38h7ae7562_2
cython-blis=0.4.1=py38he774522_1
docutils=0.15.2=py38_0
en-core-web-sm=2.3.0=pypi_0
fr-core-news-sm=2.3.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=py_0
importlib-metadata=1.7.0=py38_0
importlib_metadata=1.7.0=0
intel-openmp=2020.2=254
jmespath=0.10.0=py_0
joblib=0.16.0=py_0
jsonschema=3.2.0=py38h32f6830_1
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
murmurhash=1.0.2=py38h33f27b4_0
ninja=1.10.0=py38h7ef1ec2_0
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
openssl=1.1.1g=he774522_1
pip=20.2.2=py38_0
plac=0.9.6=py38_1
preshed=3.0.2=py38h7ae7562_3
pycparser=2.20=py_2
pyopenssl=19.1.0=py_1
pyrsistent=0.16.0=py38he774522_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.5=he1778fa_0
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda102_cudnn7_0
regex=2020.7.14=py38he774522_0
requests=2.24.0=py_0
s3transfer=0.3.3=py38h32f6830_1
sacremoses=0.0.43=pyh9f0ad1d_0
scipy=1.5.0=py38h9439919_0
setuptools=49.6.0=py38_0
six=1.15.0=py_0
spacy=2.3.0=py38h95a2b95_0
sqlite=3.33.0=h2a8f88b_0
srsly=1.0.2=py38ha925a31_0
thinc=7.4.1=py38h74a9793_0
tqdm=4.48.2=py_0
transformers=2.1.1=py_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wasabi=0.7.1=py_0
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py38_0
wincertstore=0.2=py38_0
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
